requirements.txt
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
flask==3.0.3
importlib-metadata==8.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.8.0
python-dotenv==1.0.1
pytz==2024.1
werkzeug==3.0.3
zipp==3.19.2
